validation

+++ OnlineRestaurant/src/java/com/epam/restaurant/controller/command/impl/DeleteCategoryCommand.java
@@ -4,12 +4,14 @@ import com.epam.restaurant.controller.command.Command;
 import com.epam.restaurant.controller.command.exception.CommandException;
 import com.epam.restaurant.controller.name.JspPageName;
 import com.epam.restaurant.entity.Category;
+import com.epam.restaurant.entity.Dish;
 import com.epam.restaurant.service.CategoryService;
 import com.epam.restaurant.service.exception.ServiceException;
 import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.List;
 
 /**
  * Delete category from category list by user with role ADMIN.
@@ -29,7 +31,14 @@ public class DeleteCategoryCommand implements Command {
             Category category = categoryService.getByName(name);
 
             if(category!=null){
-                categoryService.delete(category);
+                //if there dishes of this category
+                List<Dish> dishList = categoryService.getAllFromCategory(category.getId());
+                if(dishList!=null){
+                    result=JspPageName.ERROR_JSP;
+                    return result;
+                }else{
+                    categoryService.delete(category);
+                }
             }
         } catch (ServiceException e) {
             throw new CommandException("EditCategoryCommand Exception",e);
